Pull multi column renderer into separate component
